+++ js/models/app.model.drawing.js
@@ -25,6 +25,10 @@ app.model.drawing = {
       redo_count  : 0, // for redo states
       coords      : [], // cartesian coordinates
 
+      is_streaming: 'false',
+      streaming_id: 'false',
+      th_offset   : 0,
+
       d3_data : {
           background          : 'transparent', // transparent, #fdfaf3, #d6d2ca, #c9bb96
           stroke              : '#797977', // #797977, #948252
@@ -72,6 +76,39 @@ app.model.drawing = {
     this.on('change:paths', this.update_path_count);
     this.on('add:paths', this.update_path_count);
     this.on('remove:paths', this.update_path_count);
+  },
+  toggle_streaming: function() {
+    var self = this;
+    var streaming = this.get('is_streaming');
+
+    if (streaming == 'false') {
+      app.log("Start Streaming");
+
+      // turn streaming on
+      var sisbot = app.manager.get_model('sisbot_id');
+      sisbot._update_sisbot('start_streaming', {}, function(resp) {
+        app.log(resp);
+        if (resp.err) return app.log("Error", resp.err);
+
+        // save streaming_id for streaming calls
+        _.each(resp.resp, function(obj) {
+          if (obj && obj.streaming_id) self.set('streaming_id', obj.streaming_id);
+        });
+
+        self.set('is_streaming', 'true');
+      });
+    } else {
+      app.log("Stop Streaming");
+
+      // turn streaming off
+      var sisbot = app.manager.get_model('sisbot_id');
+      sisbot._update_sisbot('stop_streaming', {id:self.get('streaming_id')}, function(resp) {
+        app.log(resp);
+        if (resp.err) return app.log("Error", resp.err);
+
+        self.set('is_streaming', 'false');
+      });
+    }
   },
   firstR_change: function() {
     var firstR = this.get('edit.firstR');
@@ -363,6 +400,8 @@ app.model.drawing = {
     var lastR = this.get('edit.lastR');
     var mid = this.get('mid');
     var padding = this.get('padding');
+    var end_x = mid;
+    var end_y = mid;
 
     // correct start rho if changed after drawing
     if (paths.length > 0) {
@@ -439,8 +478,6 @@ app.model.drawing = {
             }
           }
         } else if (index == paths.length - 1) { // draw to end point on last path
-          var end_x = mid;
-          var end_y = mid;
           if (lastR == 1) { // draw straight out at same angle
             var new_dx = path.end.x - mid;
             var new_dy = path.end.y - mid;
@@ -473,13 +510,23 @@ app.model.drawing = {
       .attr("stroke-width", 0);
 
     // lastR
-    svg.append("rect")
-      .attr("width", 4)
-      .attr("height", 4)
-      .attr("fill", d3_data.end_color)
-      .attr("x", mid - 2 - (mid - padding)*lastR)
-      .attr("y", mid-2)
-      .attr("stroke-width", 0);
+    if (lastR == 0) {
+      svg.append("rect")
+        .attr("width", 4)
+        .attr("height", 4)
+        .attr("fill", d3_data.end_color)
+        .attr("x", mid-2)
+        .attr("y", mid-2)
+        .attr("stroke-width", 0);
+    } else {
+      svg.append("rect")
+        .attr("width", 4)
+        .attr("height", 4)
+        .attr("fill", d3_data.end_color)
+        .attr("x", end_x-2)
+        .attr("y", end_y-2)
+        .attr("stroke-width", 0);
+    }
 
     // start point
     if (this.get('is_dragging') == 'false') {
@@ -781,6 +828,41 @@ app.model.drawing = {
             self._line(svg, {x1:new_point[0],y1:new_point[1],x2:new_curr[0],y2:new_curr[1],stroke:d3_data.mirror_stroke,stroke_width:d3_data.stroke_width});
           }
 
+          if (this.get('is_streaming') == 'true') {
+            var mid = w/2;
+            var x = (mid - point[0])/(mid-padding);
+            var y = (mid - point[1])/(mid-padding);
+            var theta = Math.atan2(y,x);
+            var rho = Math.sqrt(x*x+y*y);
+
+            // compare to last point, see if we need to wrap theta
+            if (old_x && old_y) {
+              var th_offset = this.get('th_offset');
+          		var pi = Math.PI;
+          		var loop_th = pi * 2;
+              var x_old = (mid - old_x)/(mid-padding);
+              var y_old = (mid - old_y)/(mid-padding);
+              var theta_old = Math.atan2(y_old, x_old);
+
+              if (theta - theta_old > pi) {
+                th_offset -= loop_th;
+                this.set('th_offset', th_offset);
+        			} else if (theta - theta_old < -pi) {
+                th_offset += loop_th;
+                this.set('th_offset', th_offset);
+              }
+              theta = theta + th_offset;
+            }
+
+            // app.log("Line", point[0]-padding,point[1]-padding,old_x-padding,old_y-padding, this.get('width')-padding*2, this.get('height')-padding*2);
+            // app.log("Theta/Rho", theta, rho);
+
+            // send to table
+            var sisbot = app.manager.get_model('sisbot_id');
+            sisbot._update_sisbot('add_verts_streaming', {id:self.get('streaming_id'), verts:[{th:theta,r:rho}]}, function(obj) {
+              app.log(obj);
+            });
+          }
           self._line(svg, {x1:point[0],y1:point[1],x2:old_x,y2:old_y,stroke:d3_data.stroke,stroke_width:d3_data.stroke_width});
         }
 
@@ -942,6 +1024,14 @@ app.model.drawing = {
     this.set('drag_pos.current.y', this.get('mid'));
 
     // app.log('Current', this.get('drag_pos.current'))
+
+    if (this.get('is_streaming') == 'true') {
+      // turn streaming off
+      var sisbot = app.manager.get_model('sisbot_id');
+      sisbot._update_sisbot('clear_verts_streaming', {id: this.get('streaming_id')}, function(obj) {
+        app.log(obj);
+      });
+    }
 
     if (this.get('el_id') != 'false') {
       this._draw_preview({el_id: this.get('el_id')});